Dnd working with styles
