+++ .storybook/webpack.config.js
@@ -2,16 +2,12 @@ const path = require("path");
 const SRC_PATH = path.join(__dirname, '../packages');
 module.exports = ({config}) => {
   config.module.rules.push({
-    test: /\.(ts|tsx)$/,
+    test: /\.tsx?$/,
+    loader: "ts-loader",
+    options: {
+      configFile: path.resolve(__dirname, "../.typescript/tsconfig.json")
+    },
     include: [SRC_PATH],
-      use: [
-        {
-          loader: require.resolve('awesome-typescript-loader'),
-          options: {
-            configFileName: './.typescript/tsconfig.json'
-          }
-        }
-      ]
   }, 
   {
     test: /\.css$/,
@@ -19,7 +15,13 @@ module.exports = ({config}) => {
         path.resolve(__dirname, "not_exist_path")
     ],
     loader: "style!css"
-});
+  },
+  {
+    test: /\.scss$/,
+    use: ['style-loader', 'css-loader', 'sass-loader'],
+    include: path.resolve(__dirname, '../'),
+  }
+  );
   config.resolve.extensions.push('.ts', '.tsx');
   return config;
 };
