+++ package.json
@@ -86,6 +86,7 @@
     "@types/jest": "^24.0.15",
     "@types/lodash.differencewith": "^4.5.6",
     "@types/lodash.isequal": "^4.5.5",
+    "@types/react-addons-shallow-compare": "^0.14.22",
     "@types/react-dom": "^16.8.5",
     "@types/react-transition-group": "^4.2.3",
     "@types/storybook__react": "^4.0.2",
@@ -113,6 +114,7 @@
     "sass-loader": "^7.1.0",
     "storybook-readme": "^5.0.3",
     "ts-jest": "^24.0.2",
+    "ts-loader": "^6.2.1",
     "tslint": "^5.18.0",
     "tslint-config-prettier": "^1.18.0",
     "typescript": "^3.5.2"
