+++ packages/Blaze-ui/.npmignore
@@ -1,0 +1,4 @@
+docs/
+tests/
+example.env
+.env
