+++ packages/Blaze-ui/src/base/_base.scss
@@ -1,0 +1,55 @@
+body {
+  margin: 0;
+  padding: 0;
+  text-align: center;
+  height: 100%;
+  color: $nearly-black;
+  font-family: $primary-font;
+  background-color: rgba($blue-grey-50, .5);
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+a {
+  color: $nearly-black;
+  text-decoration: none;
+  transition: $animate-default;
+  /*border-bottom: 1px solid transparent;
+  &:hover {
+    border-bottom: 1px solid $grey;
+  }*/
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+* {
+  box-sizing: border-box;
+}
+
+img {
+  max-width: 100%;
+  height: auto;
+}
+
+legend {
+  display: none;
+}
+
+fieldset {
+  border: 0;
+  padding: 0;
+  margin: 0;
+}
+
+h2,
+h3,
+h4,
+h5,
+h6 {
+  margin: 0 0 20px;
+}
+
+p {
+  margin: 0 0 20px;
+}
