+++ packages/Blaze-ui/src/layout/_footer.scss
@@ -1,0 +1,17 @@
+.footer {
+  &--form {
+    padding: 30px;
+    margin: 20px -15px -15px;
+    background-color: $grey-100;
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    @include mq($from: phablet) {
+      margin: 20px -30px -30px;
+      padding: 30px 50px;
+    }
+    & button + button {
+      margin-left: 10px;
+    }
+  }
+}
