+++ packages/Blaze-ui/src/layout/_site.scss
@@ -1,0 +1,161 @@
+body {
+  padding-top: 50px;
+}
+
+.site-container {
+  width: 100%;
+  height: 100%;
+  overflow: scroll;
+
+  &--template {
+    padding: 10px;
+  }
+}
+
+.main-content {
+  &--login {
+    display: grid;
+    background-color: $primary-color;
+    background-image: url('https://images.thisisblaze.com/login-tile-504-432.png');
+    height: 100%;
+  }
+}
+
+.page {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+
+  &__content {
+    flex: 1 1 auto;
+    display: flex;
+    flex-direction: column;
+
+    &--with-side-panel {
+      display: flex;
+      flex-direction: column;
+
+      @include mq($from: tablet) {
+        display: grid;
+        grid-template-columns: 160px auto;
+      }
+    }
+
+    &--prompt {
+      justify-content: center;
+      display: flex;
+
+      @include mq($from: tablet) {
+        align-items: center;
+      }
+    }
+  }
+}
+
+.row {
+  display: flex;
+  align-items: center;
+  margin-bottom: 20px;
+  justify-content: space-between;
+
+  & form {
+    display: flex;
+    align-items: center;
+
+    & div {
+      margin-right: 10px;
+    }
+  }
+
+  & .form-field {
+    margin-bottom: 0;
+  }
+
+  &__text {
+    font-size: .813rem;
+    white-space: nowrap;
+  }
+
+  &--display-row {
+    flex-direction: column;
+    align-items: flex-start;
+    @include mq($from: tablet) {
+      flex-direction: row;
+    }
+  }
+}
+
+.column {
+  width: 100%;
+
+  &--one-quarter {
+    @include mq($from: tablet) {
+      max-width: 23%;
+    }
+  }
+
+  &--one-third {
+    @include mq($from: tablet) {
+      max-width: 31.5%;
+    }
+  }
+
+  &--half {
+    @include mq($from: tablet) {
+      max-width: 48.75%;
+    }
+  }
+
+  &--two-thirds {
+    @include mq($from: tablet) {
+      max-width: 66%;
+    }
+  }
+
+  &--three-quarters {
+    @include mq($from: tablet) {
+      max-width: 74.5%;
+    }
+  }
+}
+
+.content {
+  background-color: $white;
+  border: 1px solid $grey-light;
+  padding: 15px;
+  text-align: left;
+  flex: 1 1 auto;
+
+  @include mq($from: phablet) {
+    padding: 20px;
+  }
+
+  @include mq($from: tablet) {
+    padding: 30px;
+    box-shadow: $box-shadow-solid;
+  }
+
+  &__wrapper {
+    padding: 10px;
+    display: flex;
+    flex-direction: column;
+
+    @include mq($from: phablet) {
+      padding: 20px;
+    }
+
+    @include mq($from: desktop) {
+      flex-direction: row;
+    }
+  }
+}
+
+.hidden {
+  /*hidden but accesible*/
+  position: absolute !important;
+  clip: rect(1px, 1px, 1px, 1px);
+  height: 1px;
+  width: 0;
+  overflow: hidden;
+  opacity: 0;
+}
