+++ packages/Blaze-ui/src/module/_checkbox-radio.scss
@@ -1,0 +1,82 @@
+.form-radio {
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  appearance: none;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+  background-color: $white;
+  color: $primary-color;
+  border: 2px solid $primary-color;
+  height: 20px;
+  width: 20px;
+  border-radius: 20px;
+  cursor: pointer;
+  margin-right: 7px;
+  outline: none;
+  overflow: hidden;
+  transition: $animate-default;
+}
+
+.form-radio:checked::before {
+  content: '';
+  width: 10px;
+  height: 10px;
+  border-radius: 5px;
+  background-color: $primary-color;
+
+}
+.form-radio--check:checked::before {
+  position: absolute;
+  font: 600 13px/1 Arial, Helvetica, sans-serif;
+  left: 4px;
+  top: 0;
+  content: '\02143';
+  transform: rotate(40deg);
+  width: auto;
+  height: auto;
+}
+
+.form-radio:hover {
+  background-color: lighten($primary-color, 50%);
+}
+
+.form-radio:checked {
+  background-color: $white;
+}
+
+.form-checkbox {
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  appearance: none;
+  display: inline-block;
+  position: relative;
+  background-color: $white;
+  color: $white;
+  height: 20px;
+  width: 20px;
+  border: 2px solid $primary-color;
+  border-radius: 3px;
+  cursor: pointer;
+  margin-right: 7px;
+  outline: none;
+  transition: $animate-default;
+}
+
+.form-checkbox:checked::before {
+  position: absolute;
+  font: 600 13px/1 Arial, Helvetica, sans-serif;
+  left: 4px;
+  top: 0;
+  content: '\02143';
+  transform: rotate(40deg);
+}
+
+.form-checkbox:hover {
+  background-color: lighten($primary-color, 50%);
+}
+
+.form-checkbox:checked {
+  background-color: $primary-color;
+}
