+++ packages/Blaze-ui/src/module/_content.scss
@@ -1,0 +1,34 @@
+p {
+  line-height: 1.5em;
+  margin: 0 0 30px;
+  font-weight: $font-regular;
+}
+
+.modal {
+  & p {
+    &:last-of-type {
+      margin: 0;
+    }
+  }
+}
+
+.heading {
+  color: $nearly-black;
+
+  &--content {
+    margin: 0 0 20px;
+    font-size: 1rem;
+    font-weight: 600;
+  }
+}
+
+.header--list {
+  & h2 {
+    font-size: .875rem;
+    font-weight: $font-semi-bold;
+    letter-spacing: $x-small-space;
+    margin: 0;
+    white-space: nowrap;
+    color: $grey-dark;
+  }
+}
