+++ packages/Blaze-ui/src/module/_more-menu.scss
@@ -1,0 +1,129 @@
+.more-menu {
+  @extend %list;
+  position: absolute;
+  top: 38px;
+  right: 0;
+  background-color: $white;
+  border: 1px solid $grey-200;
+  box-shadow: $box-shadow-small;
+  width: 160px;
+  display: none;
+  z-index: 110;
+
+  &--open {
+    display: block;
+  }
+
+  &--header {
+    font-size: $size-x-small;
+    font-weight: 400;
+    width: 100%;
+    max-width: 300px;
+    letter-spacing: $x-small-space;
+    display: none;
+    border-color: darken($primary-color, 5%);
+  }
+  &--media {
+    width: 100%;
+    border-color: $tertiary-color;
+  }
+
+  &__background {
+    background: rgba(0, 0, 0, 0.1);
+    width: 100%;
+    height: 100%;
+    bottom: 0;
+    left: 0;
+    position: absolute;
+    right: 0;
+    top: 0;
+    z-index: 105;
+  }
+
+  &__button {
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    width: 100%;
+    color: $white;
+    padding: 5px;
+    border: 0;
+    background: $primary-color;
+    font-weight: 700;
+    cursor: pointer;
+    font-size: $size-x-small;
+
+    & i {
+      color: $amber;
+    }
+  }
+
+  &__list-item {
+    width: 100%;
+    font-size: 0.875rem;
+    border-bottom: 1px solid $grey-100;
+
+    &:last-of-type {
+      border: 0;
+    }
+
+    &--header {
+      background: $primary-color;
+      border-color: darken($primary-color, 5%);
+    }
+  }
+
+  &__link {
+    display: inline-block;
+    width: 100%;
+    padding: 12px 20px;
+    text-align: left;
+    color: $nearly-black;
+    text-decoration: none;
+    cursor: pointer;
+    border: 0;
+    background: $white;
+
+    &:hover {
+      background-color: $grey-100;
+      text-decoration: none;
+    }
+
+    &--header {
+      background: $primary-color;
+      border: 0;
+      color: $white;
+      padding: 12px;
+      text-align: left;
+      width: 100%;
+      cursor: pointer;
+      font-weight: 600;
+      font-size: $size-x-small;
+
+      &:hover {
+        background: darken($primary-color, 5%);
+      }
+    }
+  }
+
+  &__name {
+    width: 100%;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
+
+  &__wrapper {
+    position: relative;
+
+    &--header {
+      display: none;
+      @include mq($from: tablet) {
+        display: flex;
+        justify-content: flex-end;
+        width: 100%;
+        max-width: 160px;
+      }
+    }
+  }
+}
