+++ packages/Blaze-ui/src/module/_nestable.scss
@@ -1,0 +1,74 @@
+.nestable {
+  position: relative;
+}
+
+.nestable-list {
+  list-style-type: none;
+  padding: 0;
+  width: 100%;
+}
+
+.nestable .nestable-list {
+  margin: 0;
+  list-style-type: none;
+}
+.nestable > .nestable-list {
+  padding: 0;
+}
+.nestable-item,
+.nestable-item-copy {
+  margin: 10px 0 0;
+}
+.nestable-item:first-child,
+.nestable-item-copy:first-child {
+  margin-top: 0;
+}
+.nestable-item .nestable-list,
+.nestable-item-copy .nestable-list {
+  margin-top: 10px;
+}
+
+.nestable-item-copy.is-dragging {
+  opacity: 0;
+}
+
+.nestable-item {
+  position: relative;
+}
+.nestable-item.is-dragging .nestable-list {
+  pointer-events: none;
+}
+.nestable-item.is-dragging * {
+  opacity: 0;
+  -ms-filter: "progid:DXImageTransform.Microsoft.Alpha(Opacity=0)";
+  filter: alpha(opacity=0);
+}
+.nestable-item.is-dragging:before {
+  content: " ";
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: #87ceeb;
+  border: 1px dashed #4682b4;
+  -webkit-border-radius: 5px;
+  border-radius: 5px;
+}
+.nestable-item-icon {
+  margin-right: 5px;
+  cursor: pointer;
+}
+.nestable-drag-layer {
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 100;
+  pointer-events: none;
+}
+.nestable-drag-layer > .nestable-list {
+  position: absolute;
+  top: 0;
+  left: 0;
+  padding: 0;
+}
