+++ packages/Blaze-ui/src/module/_page-items.scss
@@ -1,0 +1,337 @@
+a {
+  color: $link-color;
+
+  &:hover {
+    border-color: $link-color;
+  }
+}
+
+hr {
+  margin: 20px 0;
+  padding: 0;
+  width: 100%;
+  height: 1px;
+  border: 0;
+  background-color: $grey-dark;
+}
+
+.author {
+  margin: 0 0 20px;
+}
+
+.date {
+  margin: 0 0 20px;
+}
+
+.avatar {
+  width: 32px;
+  height: 32px;
+  border-radius: 16px;
+  background-color: rgba($black, 0.2);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  line-height: 0;
+  overflow: hidden;
+  flex: 0 0 32px;
+  @include mq($from: tablet) {
+    padding: 0;
+    margin: 0 0 0 5px;
+  }
+
+  & img {
+    width: 28px;
+    height: 28px;
+    border-radius: 14px;
+  }
+}
+
+.page-header {
+  padding: 0 15px 10px;
+  background-color: $blue-grey-50;
+  border-bottom: 5px solid rgba($blue-grey-700, 0.2);
+  display: flex;
+  text-align: left;
+  @include mq($from: mobileBig) {
+    padding: 0 20px 10px;
+  }
+  @include mq($from: tablet) {
+    padding: 0 30px 10px;
+  }
+
+  &--sticky {
+    justify-content: space-between;
+    align-items: flex-start;
+    padding: 5px 15px;
+    position: sticky;
+    top: 0;
+    z-index: 100;
+    border: 0;
+    @include mq($from: mobileBig) {
+      padding: 10px 20px;
+    }
+    @include mq($from: tablet) {
+      padding: 10px 30px;
+    }
+  }
+
+  &__heading {
+    font-size: 1.125rem;
+    font-weight: $font-bold;
+    width: 100%;
+    overflow: hidden;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    @include mq($from: tablet) {
+      font-size: 1.5rem;
+    }
+  }
+
+  &__subheading {
+    font-size: .813rem;
+    color: $grey-dark;
+  }
+
+  &__button {
+    flex: 1 0 24px;
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    margin-left: 5px;
+    position: relative;
+
+    & .button {
+      margin: 0 5px;
+    }
+  }
+}
+
+.add,
+.upload {
+  padding: 20px;
+  background-color: $grey-lightest;
+  border: 2px dashed $blue-grey-100;
+  border-radius: 5px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  color: $grey-dark;
+  @include mq($from: tablet) {
+    padding: 30px;
+  }
+
+  &__drag-drop {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 0 0 30px;
+    text-align: center;
+  }
+
+  &__intro {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 0;
+    text-align: center;
+
+    & h2 {
+      font-size: 1.125rem;
+      font-weight: $font-semi-bold;
+      margin: 10px 0;
+    }
+
+    & p {
+      margin-bottom: 20px;
+    }
+  }
+
+  &__icon {
+    color: $pink;
+    height: 48px;
+    width: 48px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border: 2px solid $blue-grey-100;
+    border-radius: 24px;
+    margin-bottom: 10px;
+  }
+
+  &__browse {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 0 0 30px;
+  }
+
+  &__text {
+    margin: 0 0 20px;
+    color: lighten($grey-dark, 30%);
+  }
+}
+
+.preview {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  padding-top: 30px;
+  border-top: 2px dashed $blue-grey-100;
+  @include mq($from: tablet) {
+    flex-direction: row;
+    justify-content: space-between;
+  }
+
+  & + .preview {
+    margin-top: 30px;
+  }
+
+  &__file {
+    flex: 0 0 100%;
+    text-align: center;
+    margin-bottom: 15px;
+    border: 10px solid white;
+    box-shadow: 0 2px 2px 0 rgba(0, 0, 0, .2);
+    @include mq($from: tablet) {
+      flex: 0 1 200px;
+      margin-right: 30px;
+      margin-bottom: auto;
+    }
+
+    & img {
+      width: 100%;
+      height: auto;
+    }
+
+    & svg {
+      max-width: 50px;
+      fill: $grey;
+    }
+
+    &--doc {
+      flex: 0 1 200px;
+      border: 2px solid $blue-grey-100;
+      background: rgba($white, .5);
+      box-shadow: none;
+    }
+
+    &--video {
+      flex: 0 1 200px;
+      border: 2px solid $blue-grey-100;
+      background: rgba($white, .5);
+      box-shadow: none;
+    }
+  }
+
+  &__filename {
+    font-size: .875rem;
+    margin-bottom: 15px;
+    line-height: 1rem;
+    font-weight: 600;
+  }
+
+  &__details {
+    display: flex;
+    flex-direction: column;
+    flex: 1 1 auto;
+  }
+
+  &__form {
+    & .form-field {
+      margin-bottom: 10px;
+    }
+
+    & .button {
+      margin-top: 20px;
+    }
+  }
+}
+
+.add {
+  margin: 20px 0;
+}
+/*Background colours*/
+.red {
+  background-color: $red;
+}
+
+.pink {
+  background-color: $pink;
+}
+
+.pink-dark {
+  background-color: $pink-800;
+}
+
+.purple {
+  background-color: $purple;
+}
+
+.deep-purple {
+  background-color: $deep-purple;
+}
+
+.indigo {
+  background-color: $indigo;
+}
+
+.blue {
+  background-color: $blue;
+}
+
+.light-blue {
+  background-color: $light-blue;
+}
+
+.cyan {
+  background-color: $cyan;
+}
+
+.teal {
+  background-color: $teal;
+}
+
+.green {
+  background-color: $green;
+}
+
+.light-green {
+  background-color: $light-green;
+}
+
+.lime {
+  background-color: $lime;
+}
+
+.yellow {
+  background-color: $yellow;
+}
+
+.amber {
+  background-color: $amber;
+}
+
+.orange {
+  background-color: $orange;
+}
+
+.deep-orange {
+  background-color: $deep-orange;
+}
+
+.brown {
+  background-color: $brown;
+}
+
+.grey-dark {
+  background-color: $grey-dark;
+}
+
+.purple-300 {
+  background-color: $purple-300;
+}
+
+.purple-800 {
+  background-color: $purple-800;
+}
