+++ packages/Blaze-ui/src/module/_sidebar.scss
@@ -1,0 +1,22 @@
+.sidebar {
+  margin-top: 10px;
+  @include mq($from: desktop) {
+    flex: 0 1 210px;
+    margin-top: 0;
+    margin-left: 30px;
+  }
+
+  &__content {
+    background-color: $blue-grey-50;
+    padding: 20px;
+    border-bottom: 5px solid rgba($blue-grey-700, .2);
+  }
+
+  & .form-field {
+    margin: 0 0 20px;
+
+    &:last-of-type {
+      margin: 0;
+    }
+  }
+}
