+++ packages/Blaze-ui/src/module/_switches.scss
@@ -1,0 +1,118 @@
+.switch {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  margin: 0 0 30px;
+
+  & input[type="checkbox"] {
+    height: 0;
+    width: 0;
+    display: none;
+  }
+
+  & label {
+    cursor: pointer;
+    text-indent: -9999px;
+    width: 40px;
+    height: 18px;
+    background: $grey;
+    display: block;
+    border-radius: 20px;
+    position: relative;
+    margin: 0;
+
+    &:after {
+      content: "";
+      position: absolute;
+      top: 2px;
+      left: 3px;
+      width: 14px;
+      height: 14px;
+      background: $white;
+      border-radius: 18px;
+      transition: all 0.3s;
+    }
+
+    &:active:after {
+      width: 40px;
+    }
+  }
+
+  & input:checked + label {
+    background: $grey;
+
+    &:after {
+      left: calc(100% - 3px);
+      transform: translateX(-100%);
+      background: $grey-dark;
+    }
+  }
+
+  &--primary {
+    & input:checked + label {
+      background: lighten($primary-color, 30%);
+
+      &:after {
+        background: $primary-color;
+      }
+    }
+  }
+
+  &--secondary {
+    & input:checked + label {
+      background: lighten($secondary-color, 30%);
+
+      &:after {
+        background: $secondary-color;
+      }
+    }
+  }
+
+  &--disabled {
+    opacity: 0.5;
+    cursor: default;
+    pointer-events: none;
+
+    & label {
+      &:after {
+        background-color: $grey-dark;
+      }
+    }
+  }
+
+  &--label-left {
+    flex-direction: row;
+  }
+
+  &--label-right {
+    flex-direction: row-reverse;
+    justify-content: flex-end;
+  }
+
+  &--label-top {
+    flex-direction: column;
+  }
+
+  &--label-base {
+    flex-direction: column-reverse;
+  }
+
+  & + .switch {
+    margin-left: 10px;
+  }
+
+  &__wrapper {
+    display: flex;
+  }
+
+  &__item {
+    display: flex;
+    padding: 5px;
+  }
+
+  &__text {
+    font-weight: 400;
+    font-size: 0.875rem;
+    white-space: nowrap;
+  }
+}
