+++ packages/Blaze-ui/src/module/_tabs.scss
@@ -1,0 +1,67 @@
+.tabs {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  &--modal {
+    margin: -30px 0;
+  }
+
+  &__list {
+    margin: 0;
+    padding: 0;
+    list-style: none;
+    display: flex;
+    width: 100%;
+  }
+
+  &__list-item {
+    cursor: pointer;
+    display: flex;
+    flex: 1 1 auto;
+    justify-content: center;
+    padding: 10px;
+    font-size: .875rem;
+    font-weight: $font-semi-bold;
+    color: $grey-medium;
+    border: 0;
+    background: none;
+    border-bottom: 5px solid $white;
+
+    &.current {
+      border-color: $secondary-color;
+      color: $grey-dark;
+    }
+    @include mq($from: phablet) {
+      padding: 10px 30px;
+    }
+  }
+
+  &__content {
+    width: 100%;
+    display: none;
+    justify-content: space-between;
+    border-top: 1px solid $grey-light;
+    padding: 30px 0;
+
+    &.current {
+      display: flex;
+      flex-direction: column;
+    }
+
+    &--modal {
+      padding-bottom: 0;
+    }
+  }
+}
+/*Below styles tabs in a modal until a modifier can be added to tabs & tab__content if in modal. Classes for that already added above & these can be deleted then*/
+.modal__content {
+  & .tabs {
+    margin: -30px 0;
+  }
+
+  & .tabs__content {
+    padding-bottom: 0;
+  }
+}
