+++ packages/Blaze-ui/src/module/_tags.scss
@@ -1,0 +1,7 @@
+.tags {
+  @extend %list;
+  &__tags-item {
+    display: inline-block;
+    margin-right: 5px;
+  }
+}
