+++ packages/Blaze-ui/src/module/_tooltip.scss
@@ -1,0 +1,55 @@
+.tooltip {
+  top: 42px;
+  color: $nearly-black;
+  display: inline-block;
+  left: 0;
+  background-color: $white;
+  opacity: 0;
+  padding: 20px 20px 30px;
+  pointer-events: none;
+  position: absolute;
+  width: 100%;
+  z-index: 101;
+  transform: translateY(10px);
+  transition: all .25s ease-out;
+
+  &:before {
+    /* This bridges the gap so you can mouse into the tooltip without it disappearing */
+    bottom: -20px;
+    content: '';
+    display: inline-block;
+    height: 20px;
+    left: 0;
+    position: absolute;
+    width: 100%;
+  }
+
+  &__heading {
+    font-weight: 700;
+    text-transform: uppercase;
+    margin-bottom: 15px;
+    padding-right: 30px;
+  }
+
+  &__subheading {
+    font-size: .75rem;
+    text-transform: uppercase;
+    margin-bottom: 15px;
+
+    & span {
+      text-transform: none;
+      padding-left: 10px;
+    }
+  }
+
+  &__content {
+    font-size: .75rem;
+  }
+}
+
+.f-all-filters__filter:hover .tooltip,
+.table-fields__field:hover .tooltip {
+  opacity: 1;
+  pointer-events: auto;
+  transform: translateY(0px);
+}
